requirements/dev.txt
-r common.txt

django-debug-toolbar==1.2.2
Fabric==1.10.0
